feat(protocol-designer): warn user when exporting w/o steps
